Optimizations and some new features
Now the bot won't consume as much resources as it used to. Aside from that, it now also allows the user to pause it by clicking on the chat input, choose a template file other than "template.png".

Code was cleaned a bit and the service worker now properly handles debugger attachment.

Co-Authored-By: Danyil <142193142+Vaulco@users.noreply.github.com>